Adicionando a tela de Visualizaçãodo Carrinho, global_json

+++ LojaShopping/LojaShopping.Web/Models/CartHeaderViewModel.cs
@@ -5,6 +5,6 @@ public class CartHeaderViewModel
    public long Id { get; set; }
    public string UserId { get; set; }
    public string CouponCode {  get; set; }
-   public Double ValorFinal {  get; set; }
+   public Decimal ValorFinal {  get; set; }
 
 }
